refactor: Move sliders into their own file
